requirements.txt
requests
pyyaml
tqdm
pyyaml
aiohttp